css/css-tables/width-distribution/colspan_redistribution.tentative.html

<!doctype html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<link rel="author" title="Aleks Totic" href="atotic@chromium.org" />
<link rel="help" href="https://www.w3.org/TR/css-tables-3/#distributing-width-to-columns" />
<style>
  table {
    background: gray;
    border-spacing: 8px 8px;
  }
  td {
    padding: 0;
    background: #BFB;
    font-size: 10px;
  }
  td > div {
    display: inline-block;
    background: rgba(56,162,56,0.3);
  }
  .error {
    color: red;
  }
  p {
    margin-top:4px;
    margin-bottom:4px;
  }
  .rule1 {
    background: #87dc8a;
  }
  .rule2 {
    background: #3ae4cc;
  }
  .rule1and2 {
    background: #fda4a4;
  }
  body {
    counter-reset: testid;      /* Create a chapter counter scope */
  }
  .testdesc {
    white-space: pre-line;
    margin-top: 16px;
  }
  .testdesc:before {
    content: counter(testid) ". ";
    counter-increment: testid;  /* Add 1 to chapter */
  }
  .testdesc:first-line {
    font-family: monospace;
  }
</style>
<h1>Colspan > 1 width redistribution</h1>
<p>Principles</p>
<ul>
  <li>Distribution should be stable: reordered span cells retain their width.</li>
</ul>
<p>Wide cell redistributes its different widths to the spanned cells: percentage, min and max.</p>
<p>Creating understandable tests with percentage cells is complex because the relationship between percentage cell width, and table width is complex. Rules that govern relationship between table grid width and percentage cell width are:</p>
<ol>

  <li id="rule1"><span  class="rule1">Rule#1</span>, Single column sets smallest maximum table width. That table width is column.min_width / column.percent * 100.</li>
  <li id="rule2"><span class="rule2">Rule#2</span>, All percentage columns set smallest maximum table widths. <br>Let P% be sum of all percentages, and Fpx sum of widths of all non-percentage columns. The equations below determine table width:
    <ul>
    <li>table width = Mpx + Ppx </li>
    <li>P% + F% = 100%</li>
    <li>Ppx * P% = Fpx * F% </li>
    <li>Ppx * P% = Fpx * (100-F%) </li>
    <li>Ppx = Fpx * (100-P%) / P% </li>
    <li>table width = Fpx + Fpx * (100-P%) / P%.</li>
    <li>table width = Fpx * (1 + (100-P%)/P%)</li>
    </ul>
  </li>
  <p>This is a little confusing, rule 2 does not apply to minimum table width, just the maximum.</p>
  <p>Table backgrounds will change if they match the 2 rules above: <span class="rule1">Rule#1</span>, <span class="rule2">Rule#2</span>, and both <span class="rule1and2">Rule#1 and Rule#2</span>.</p>
</ol>
<h2>Test design</h2>
<p>All examples have border-spacing:8, td.padding:0.</p>
<p>Tests are tables with 2 rows. 1st row are the columns, 2nd row is the wide column being distributed. Most test's wide cell colspan encloses all cells except the last one. Each test is accompanied by test description. Test description contains:</p>
<ol>
  <li>First line describes the test geometry: cell:css/min/max. C1:20%/20px means width:20%, intrinsic min width 20px. If max is omitted, min == max.</li>
  <li>Second line describes algorithm being tested in detail</li>
  <li>Third line describes computation that generates the result.</li>
  <li class="error">Red paragraphs are major browser disagreements</li>
</ol>

<h2>Wide cell's percentage distribution</h2>
<p>Rules</p>
<ul>
  <li>Percentages can only be redistributed to non-percentage cells.</li>
  <li>If percentage does not get redistributed, treat wide cell width as Auto</li>
  <li>If all columns are empty (no max width), redistribute percentage evenly.</li>
</ul>

<p class="testdesc">C0:10%/20px C1:10%/20px C2:auto
No wide cells, shows what table looks like without wide cell distribution.
Table width by rule#1, 20px/0.1(10%) + 4*8 => 232px. Excess table width is distributed to auto cell.</p>
<table data-expected-width="232">
  <tr>
    <td style="width:10%"><div style="width:20px">x</div></td>
    <td style="width:10%"><div style="width:20px">x</div></td>
    <td>x</td>
  </tr>
</table>

<p class="testdesc">Wide:40%/Auto C0:10%/20px C1:10%/20px C2:auto
Percentage does not get redistributed because all columns are percentages.
Table width by rule#1 same as previous example.</p>
<table data-expected-width="232">
  <tr>
    <td style="width:10%" data-expected-width="20"><div style="width:20px">x</div></td>
    <td style="width:10%"><div style="width:20px">x</div></td>
    <td>x</td>
  </tr>
  <tr>
    <td colspan=2 style="width:40%">40%</td>
  </tr>
</table>

<p class="testdesc">Wide:20%/100px C0:Auto/0 C1:Auto/0 C2:Auto/Auto
Percentage gets redistributed evenly to empty cells.
Each cell gets 10%, then (100-8)/2=>46px min width. Table min width is 46/0.1(10%) + 4*8 => 460 + 32 => 492</p>
<p class="error">Chrome is wrong, 1st span cell gets all the width. FF/Edge agree.</p>
<table data-expected-width="492">
  <tr>
    <td data-expected-width="46"></td>
    <td></td>
    <td>x</td>
  </tr>
  <tr>
    <td colspan=2 style="width:20%"><div style="width:100px">100px</div></td>
</table>

<p class="testdesc">Wide: 19%/200px wide cell, C0: 20%.20px, C1: 80px/80px
Percentage is not getting redistributed, because column% > wide%</p>
<table data-expected-width="224">
  <tr>
    <td style="width:20%" data-expected-width="40"><div style="width:20px">20</div></td>
    <td style="width:80px" data-expected-width="160"><div style="width:80px">80</div></td>
  </tr>
  <tr>
    <td colspan=2 style="width:19%" data-expected-width="208"><div style="width:208px">208</div></td>
  </tr>
</table>

<h2>Wide cell's minimum width distribution</h2>
<h3>All unconstrained columns</h3>

<p class="testdesc">Wide: Auto/300px C0:Auto/100px C1:Auto:100px; C2:20px/Auto
Basic distribution over auto columns, cells grow evenly.
Distribute 300-8=292 proportional to maxiwdith. C0 gets 75/100*292=219.</p>
<table data-expected-width="344">
  <tr>
    <td data-expected-width="219"><div style="width:75px">100px</div></td>
    <td><div style="width:25px">25px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td data-expected-width="300" colspan=2><div style="width:300px">300px min</div></td>
  </tr>
</table>

<p class="testdesc">Wide: 260px/300px C0:Auto/100px C1:Auto/100px C2:Auto/Auto
Wide cell min width wins over css width.
Table width same as previous example.</p>
<table data-expected-width="344">
  <tr>
    <td data-expected-width="146"><div style="width:100px">100px</div></td>
    <td><div style="width:100px">100px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td data-expected-width="300" colspan=2 style="width:260px"><div style="width:300px">300px min</div></td>
  </tr>
</table>

<h3>All fixed columns</h3>
<li>column gets min_width proportional to its max width.

<p 7 class="testdesc">Wide: 260/300 C0:100/50/100 C1:100/100 C2:20/Auto
Distribute surplus proportional to max width.
Wide is distributing 292, surplus 92, each column gets 46.
</p>
<table data-expected-width="344">
  <tr>
    <td style="width:100px" data-expected-width="146"><div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td style="width:100px"><div style="width:100px">100px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td data-expected-width="300" colspan=2 style="width:260px"><div style="width:300px">300px min</div></td>
  </tr>
</table>

<p 8 class="testdesc">Wide: 300/300 C0:100/25 C1:100/75 C2:20/Auto
Column minimum width is less than its CSS width.
Just like last example, both columns get 46.
<table data-expected-width="344" style="width:1">
  <tr>
    <td style="width:100px" data-expected-width="146"><div style="width:25px">25px</div></td>
    <td style="width:100px"><div style="width:75px">75px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td data-expected-width="300" colspan=2 style=""><div style="width:300px">300px min</div></td>
  </tr>
</table>

<p class="testdesc">Wide: 300/300 C0:20/100 C1:100/75 C2:20/Auto
First column's min width > css width.
Like last example, each column minimum width increases in proportion to the max.</p>
<p class="error">Chrome differs from FF/Edge. It distributes min-width in proportion to css width, not max width</p>
<table data-expected-width="344">
  <tr>
    <td style="width:20px" data-expected-width="146"><div style="width:100px">20/100px</div></td>
    <td style="width:40px"><div style="width:100px">40/100px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td data-expected-width="300" colspan=2 style=""><div style="width:300px">300px min</div></td>
  </tr>
</table>

<h3>All percent columns</h3>
  <ul>
    <li>min width never gets smaller than it started.
    <li>min width becomes cell.percent/cells.percent * wide_cell.min_width</li>
    <li>max width becomes cell.percent/cells.percent * wide_cell.max_width</li>
  </ul>

<p 10 class="testdesc">Wide:Auto/300px C0:25%/50 C1:25%/30 C2:20/Auto
Wide min width gets distributed proportional to percentage (not min/max width).
Columns get min_width = 146, causes table width of 146/0.25 + 4*8 = 616.
<table data-expected-width="616">
  <tr>
    <td style="width:25%" data-expected-width="146"><div style="width:50px">25%/50</div></td>
    <td style="width:25%"><div style="width:30px">25%/30</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td colspan=2 style=""><div style="width:300px">300px min</div></td>
  </tr>
</table>

<p 11 class="testdesc">Wide:Auto/400px C0:20%/50 C1:60%/50 C2:20/Auto.
Wide min width gets distributed proportional to percentage (not min/max width).
Cells get 98/294, table by rule#1 98px/0.2 + 32 = 522</p>
<p class="error">Edge disagrees, table is 870</p>
<div style="width:600px">
<table data-expected-width="522">
  <tr>
    <td style="width:20%" data-expected-width="98"><div style="width:50px">20%/50px</div></td>
    <td style="width:60%" data-expected-width="294"><div style="width:50px">60%/50px</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td colspan=2 style=""><div style="width:400px">400px min</div></td>
  </tr>
</table>
</div>

<p 12 class="testdesc">Wide:Auto/400 C0:50%/150 C1:30%/150 C2:20/Auto
This tests conflict resolution where min-width > redistributed min width.
400-8px distributed max width tries to redistribute as 245|147, but gets constrained to 245|150 in Chrome.
table width from cell 1 245/0.5 + 4*8 = 522
table width from cell 2 150/0.3 + 4*8 = 532>
cell 1 = 50% of 500 = 250, cell 2 = 30% of 500 = 150 , cell 3 gets the remaining 100</p>
<p class="error">Chrome/FF/Edge end up with tables of different widths: 532/590/685. Chrome's 2nd span cell seems 'most correct' at its original max width of 150. In FF, extra min-width seems to cause more width to be redistributed. If you hover over 30%/150 cell, its min width will change to 100px, and all browsers will agree.</p>
<style>
  .test12:hover {
    width:100px !important;
  }
</style>
<table data-expected-width="532">
  <tr>
    <td style="width:50%" data-expected-width="250"><div style="width:150px">50%/150px</div></td>
    <td style="width:30%" data-expected-width="150"><div class="test12" style="width:150px">30%/150px</div></td>
    <td style="width:20px" data-expected-width="100">x</td>
  </tr>
  <tr>
    <td colspan=2 style=""><div style="width:400px">400px min</div></td>
  </tr>
</table>

<p 13 class="testdesc"> Wide:Auto/400px C0:50%/75px/125px, C1:30%/75px/125px C2:20px/Auto
Wide cell distribution over different percentages.
400-8px min width gets redistributed as 245/147 (no min width limits)
</p>
<p class="error">Edge is different, table is 685 instead of 522.</p>
  <table data-expected-width="522">
  <tr>
    <td style="width:50%" data-expected-width="245"><div style="width:75px">50%/75</div><div style="width:50px">/125</div></td>
    <td style="width:30%" data-expected-width="147"><div style="width:75px">30%/75</div><div style="width:50px">/125</div></td>
    <td style="width:20px">x</td>
  </tr>
  <tr>
    <td colspan=2 style=""><div style="width:400px">400px</div></td>
  </tr>
</table>


<h3>Percentage/fixed mix columns</h3>
<p>The Legacy code that deals with this is confusing, it tries to redistribute minimum width in proportion to max width, unless there are auto cells. All the browsers disagree on final widths. Some things everyone agrees on:
    <li>if there are auto cells, fixed cells do not grow.</li>
</p>

<p 14 class="testdesc">Wide:N/A C0:60%/100 C2:100/100
An example what table looks like without wide distribution.
Table width by rule#2: (60/(100-60) +1) * 100 = 250 + 24 = 274</p>
<table data-expected-width="274">
  <tr>
    <td data-expected-width="150" style="width:60%"><div style="width:100px">60%</div></td>
    <td data-expected-width="100" style="width:100px"><div style="width:100px">100px</div></td>
  </tr>
</table>

<p 15 class="testdesc">Wide:N/A C0:50%/100 C1:50%/100 C2:100/100
Total column percent is 100%, forcing maximum table grid width by rule #2 to infinity.
Table css width is 1px, overriding grid max width.
Final column width is column's minimum width.
</p>
<div style="width:500px">
  <table style="width:1px">
    <tr>
      <td data-expected-width="100" style="width:50%"><div style="width:100px">50%</div></td>
      <td data-expected-width="100" style="width:50%"><div style="width:100px">50%</div></td>
      <td data-expected-width="100" style="width:100px"><div style="width:100px">100px</div></td>
    </tr>
  </table>
</div>

<p 16 class="testdesc">Wide:N/A C0:50%/100 C1:50%/100 C2:100/100
Same example as above, but table width is auto.
Table grows to size of containing block.</p>
<div style="width:500px;display:block">
  <table>
    <tr>
      <td data-expected-width="184" style="width:50%"><div style="width:100px">50%</div></td>
      <td data-expected-width="184" style="width:50%"><div style="width:100px">50%</div></td>
      <td data-expected-width="100" style="width:100px"><div style="width:100px">100px</div></td>
    </tr>
  </table>
</div>

<p 17 class="testdesc">Wide: Auto/200px C0:40%/20 C1:50/50 C2:100/20
Table css width is 1px. This forces grid minimum.
Browsers all render it differently, varying widths of C0, C1.
FF algorithm looks correct, 40% column ends close to 40%
FF computation: distributable_min_width = 192 * (100%-40%) = 115.2. Distribute to cell #2.
Chrome algorithm: distributes proportionally to min_width.
C1.min = 192 * 20/70 = 54, C2.min = 192 * 50/70 = 138. 40% cell ends up being 25%.
Edge distributes all the width to the %ge cell.
Browsers agree that C2 width does not change, and that wide cell's width is 200
<p class="error">Everyone disagrees on how to distribute width between % and fixed cell.</p>
<table style="width:1px" data-expected-width="244">
  <tr>
    <td style="width:40%"><div style="width:20px">40%</div></td>
    <td style="width:50px"><div style="width:50px">50</div></td>
    <td data-expected-width="20" style="width:100px"><div style="width:20px">20</div></td>
  </tr>
  <tr>
    <td colspan=2 data-expected-width="200" ><div style="width:200px">200px min</div></td>
  </tr>
</table>

<h2>Misc Column widths and table width</h2>

<p class="testdesc">C0:10%/70, C1:Auto/50.
Tests table max width from single cell.
<table data-expected-width="324">
  <tr>
    <td style="width:20%"  data-expected-width="60"><div style="width:60px">60</div></td>
    <td><div style="width:50px">50</div></td>
  </tr>
</table>

<p class="testdesc">C0:10%/70, C1:Auto/50.
Table limited to 1px. Tests that single cell specifies max width, not min width.
<table style="width:1px" data-expected-width="134">
  <tr>
    <td style="width:20%"  data-expected-width="60"><div style="width:60px">60</div></td>
    <td><div style="width:50px" data-expected-width="50">50</div></td>
  </tr>
</table>

<p class="testdesc">C0:10%/70 border 10px, C1:Auto/50.
Cell border padding do not affect max width.
<table data-expected-width="524">
  <tr>
    <td style="width:20%;border:10px solid yellow;padding:10px"  data-expected-width="100"><div style="width:60px">60</div></td>
    <td><div style="width:50px">50</div></td>
  </tr>
</table>

<script>
// Table inspector. Annotates cells with table measures.
function measureCellMinMax(cell) {
  let d = document.createElement("div");
  let clone = cell.cloneNode(true);
  d.classList.add("measure");
  for (child of Array.from(clone.childNodes))
    d.appendChild(child);
  d.style.width = "min-content";
  document.body.appendChild(d);
  let min = d.offsetWidth;
  d.style.width = "max-content";
  let max = d.offsetWidth;
  d.remove();
  return {min: min, max: max};
}
function annotateTable(t) {
  let tableWidth = t.offsetWidth;
  let inlineBorderSpacing = parseInt(window.getComputedStyle(t).borderSpacing.split(' ')[0]);
  let spacing = (2 + t.querySelector("tr").querySelectorAll("td").length - 1)*inlineBorderSpacing;
  t.setAttribute("title", `Table: ${tableWidth.toFixed(0)} spacing: ${spacing}px`);
  let firstCell;
  let totalCellPercent = 0;
  let nonPercentCellMinWidth = 0;
  let nonPercentCellMaxWidth = 0;
  let tableMinWidthByCellPercent = 0;
  // Spacing assumes column count specified by the 1st row. Assumption might be wrong.
  let cells = Array.from(t.querySelectorAll("td"));
  for (let cell of cells) {
    if (!firstCell)
      firstCell = cell;
    let percent = cell.offsetWidth / (tableWidth - spacing) * 100;
    let minmax = measureCellMinMax(cell);
    let title = `${cell.offsetWidth.toFixed(1)}px\nmin:${minmax.min}px max:${minmax.max}px ${percent.toFixed(1)}%`;
    let cssWidth = cell.style.width;
    let is_percent = cssWidth && cssWidth.match(/\%/);
    if (is_percent) {
      let w = parseFloat(cssWidth);
      totalCellPercent += w;
      let tableMinWidth = minmax.min / (w / 100);
      tableMinWidthByCellPercent = Math.max(tableMinWidthByCellPercent, tableMinWidth);
      title += `\nmin table: ${tableMinWidth.toFixed(0)}px`;
    } else {
      nonPercentCellMinWidth += minmax.min;
      nonPercentCellMaxWidth += minmax.max;
    }
    title += `\nTable: ${tableWidth.toFixed(0)} spacing: ${spacing}px`;
    cell.setAttribute("title", title);
  }


  // Display table statistics in first cell.
  if (firstCell) {
    let title = firstCell.getAttribute("title");
    let ruleMatch = 0;
    if (tableMinWidthByCellPercent != 0) {
      if ((tableMinWidthByCellPercent + spacing) == tableWidth)
        ruleMatch += 1;
      title += `\nTable by rule 1 min : ${tableMinWidthByCellPercent.toFixed(1)}`;
    } else {
      title += "\nTable min by single cell percent NA";
    }
    if (totalCellPercent > 0) {
      totalCellPercent = Math.min(totalCellPercent, 100);
      let minByCellPercent
    }
    if (nonPercentCellMinWidth && totalCellPercent > 0) {
      totalCellPercent = Math.min(totalCellPercent, 100);
      title += `\nsum%: ${totalCellPercent.toFixed(1)}%; non% min: ${nonPercentCellMinWidth}px; non% max ${nonPercentCellMaxWidth}px`;
      let tableMinBySum = (totalCellPercent / (100 - totalCellPercent) +1) * nonPercentCellMinWidth;
      let tableMaxBySum = (totalCellPercent / (100 - totalCellPercent) +1) * nonPercentCellMaxWidth;
      if (Math.floor((tableMinBySum + spacing)) == Math.floor(tableWidth) ||
        Math.floor((tableMaxBySum + spacing)) == Math.floor(tableWidth))
        ruleMatch += 2;
      title += `\nTable by rule 2 ${totalCellPercent}%; min:${tableMinBySum.toFixed(1)}px; max:${tableMaxBySum.toFixed(1)}px;`;
    } else {
      "Table min by % sum not available";
    }
    firstCell.setAttribute("title", title);
    switch(ruleMatch) {
      case 1: t.classList.toggle('rule1'); break;
      case 2: t.classList.toggle('rule2'); break;
      case 3: t.classList.toggle('rule1and2'); break;
      default: break;
    }
  }
}
for (let t of Array.from(document.querySelectorAll("table")))
  annotateTable(t);
checkLayout("table");
</script>
